internal/lsp: Add support for hovering runes
Enable to hover runes expressed in various forms within basic literals.
A quick summary is displayed if a rune is found: a printable version (if
it exists), its codepoint and its name.

Rune literals always display the summary when hovered, string literals
only display it when an escaped rune sequence is foudn, and number
literals only when it is expressed as a hexadecimal number whose size
ranges from one to eight bytes.

Fixes golang/go#38239